refactor: merge run_command and command_msg functions
